Setup a simple Ant + Ivy project
The project uses ant to build the project and ivy to manage
dependencies. To build and run this project run ant in the project
directory.

The demo class drones.Hello depends on Jama. By running ant, you will
download the required library from maven and place a copy of the jar
file in the lib directory.

After running ant successfully, you can run the project by invoking java
on the command line like so:

java -cp "build:lib/*" drones.Hello

Note for the above command to work, make sure you are in the project
directory.